Clean up and upgrade nuget packages
There are a number of packages that we declared references on that Resharper reports as 'implicitly used', and are generally included as part of the runtime. It's better to not have the package reference if we don't need a specific feature newer than what the runtime provides, and at this point the runtime will be either .NET Framework 4.8+ or .NET 8.0+, all older versions are "out of support" in some form.